feat: Add a 404 error page on profile page
This commit adds a 404 error page for profile page when the user data is undefined or the slug is not a string. It displays "404" on the page to inform the user that the profile page was not found. A check was added to throw an error if the slug is not a string.