feat: Update Card component with nextAudioFile state
This commit updates the Card component by introducing the `nextAudioFile` state to keep track of the index of the next audio file to be played. When the `handlePlayAudio` function is called, it plays the audio file corresponding to `nextAudioFile` and updates `nextAudioFile` to the index of the next audio file in the `audioUrls` array.